+++ src/Domains/Connectors/Elead/Entities/LeadSource.php
@@ -2,17 +2,21 @@
 
 declare(strict_types=1);
 
-namespace Kanvas\Connectors\Elead\Services;
+namespace Kanvas\Connectors\Elead\Entities;
 
 use Baka\Contracts\AppInterface;
 use Generator;
 use Kanvas\Companies\Models\Companies;
 use Kanvas\Connectors\Elead\Client;
 
-class LeadStatus
+class LeadSource
 {
-    public string $status;
-    public array $subStatus = [];
+    public string $name;
+    public string $upType;
+    public bool $isActive;
+    public bool $requiresSubSource;
+    public bool $hasSubSources;
+    public array $links = [];
 
     /**
      * Assign value to the current object.
@@ -28,7 +32,7 @@ class LeadStatus
     {
         $client = new Client($app, $company);
         $response = $client->get(
-            '/sales/v1/elead/productreferencedata/companyOpportunityStatuses'
+            '/sales/v1/elead/productreferencedata/companyOpportunitySources'
         );
 
         foreach ($response['items'] as $item) {
