+++ src/Domains/Connectors/Elead/Entities/LeadStatus.php
@@ -2,21 +2,17 @@
 
 declare(strict_types=1);
 
-namespace Kanvas\Connectors\Elead\Services;
+namespace Kanvas\Connectors\Elead\Entities;
 
 use Baka\Contracts\AppInterface;
 use Generator;
 use Kanvas\Companies\Models\Companies;
 use Kanvas\Connectors\Elead\Client;
 
-class LeadSource
+class LeadStatus
 {
-    public string $name;
-    public string $upType;
-    public bool $isActive;
-    public bool $requiresSubSource;
-    public bool $hasSubSources;
-    public array $links = [];
+    public string $status;
+    public array $subStatus = [];
 
     /**
      * Assign value to the current object.
@@ -32,7 +28,7 @@ class LeadSource
     {
         $client = new Client($app, $company);
         $response = $client->get(
-            '/sales/v1/elead/productreferencedata/companyOpportunitySources'
+            '/sales/v1/elead/productreferencedata/companyOpportunityStatuses'
         );
 
         foreach ($response['items'] as $item) {
